+++ onprem.js
@@ -39,8 +39,8 @@ var regions = ['WJU', 'PHX'];                                     // Region type
 function getSelectedObject() {
     const cloud = 'onprem';
     const region = $('#region').val();
-    const cpu = parseInt($('#cpu').val()) || 0;
-    const memory = parseInt($('#memory').val()) || 0;
+    const cpu = parseFloat($('#cpu').val()) || 0;
+    const memory = parseFloat($('#memory').val()) || 0;
     const count = parseInt($('#count').val()) || 0;
     const year = 1;
 
@@ -49,7 +49,7 @@ function getSelectedObject() {
 
 function getSelectedDiskObject() {
     const region = $('#disk_region').val();
-    const disk = parseInt($('#disk').val()) || 0;
+    const disk = parseFloat($('#disk').val()) || 0;
     const count = parseInt($('#disk_count').val()) || 0;
     const year = 1;
